pyproject.toml
[project]
name = "unsloth-qwen3-vl-peft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "xformers==0.0.32.post2",
    "bitsandbytes==0.48.2",
    "peft==0.17.1",
    "trl==0.22.2",
    "triton==3.4.0",
    "cut_cross_entropy==25.1.1",
    "sentencepiece==0.2.1",
    "unsloth_zoo==2025.11.4",
    "dataset==3.6.0",
    "transformers==4.57.0",
    "huggingface_hub>=0.34.0"
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
